core 3.6: ShapoGFX 4f15411, an opaque HUD shadow without blending
- ShapoGFX 4f15411: textMetrics() returns integers (the float one is now
  textMetricsF()), so measuring text takes no soft float on the M0+ and
  the ESP8266 (1,579 against 3,333 instructions a call on the M0+).
- With DEVOURSPHERE_SUPPRESS_ALPHA (PicoSystem, ESPboy) the drop shadow
  under HUD text is solid black instead of alpha 180: a translucent glyph
  reads the frame back per pixel (31k -> 15k instructions for 12
  characters on the M0+).

The default build renders exactly as 3.5 (fixed-run hash at 240x240 and
640x360); the suppressed build differs only in the shadows.

+++ impl/cli/src/aa.cpp
@@ -64,8 +64,8 @@ void Converter::buildTable() {
   g2::Graphics2D g(s);
   g.setFont(&ShapoSansMono_s08c07);
   g.setTextColor(g2::Colors::WHITE);
-  const int adv = (int)g.charMetrics('M').width;
-  const int lh = (int)g.charMetrics('M').height;
+  const int adv = g.charMetrics('M').width;
+  const int lh = g.charMetrics('M').height;
   uint32_t seed[95];
   for (int c = 32; c < 127; c++) {
     g.clear(g2::Colors::BLACK);
